+++ src/main/java/Episante/back/Models/Patient.java
@@ -1,30 +1,157 @@
 package Episante.back.Models;
 
 import jakarta.persistence.*;
-import jakarta.validation.constraints.Size;
-import lombok.Data;
 
-@Data
+import java.sql.Timestamp;
+
+// Cette annotation de JPA (Java Persistence API) indique que cette classe est une entité
 @Entity
 public class Patient {
+
+    // Cette annotation marque le champ id comme étant la clé primaire de l'entité
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
+
+    @Column(unique = true, nullable = false)
+
+    private String identifier;
     private String nom;
     private String prenom;
-    private String age;
-    private String adresse;
-    @Column(unique = true, nullable = false)
-    private String email;
-    @Size(min = 6, message = "Le mot de passe doit contenir au moins 6 caractères.")
-    private String mdp;
-    private String telephone;
-    private String taille;
-    private String poids;
+    private int age;
+    private double poids;
+    private double currentWeight;
+    private Double height;
     private Sexe sexe;
-    @OneToOne(mappedBy = "patient", cascade = CascadeType.ALL, orphanRemoval = true)
-    private H_Medical h_m;
+    private String email;
+    private String phoneNumber;
+    private String address;
+
+    @Column(updatable = false)
+    private Timestamp createdAt;
+
+    private Timestamp updatedAt;
+
+    @Column(nullable = false)
+    private String password; // Added password field
 
 
+    public long getId() {
+        return id;
+    }
 
+    public void setId(long id) {
+        this.id = id;
+    }
+
+    public String getNom() {
+        return nom;
+    }
+
+    public void setNom(String nom) {
+        this.nom = nom;
+    }
+
+    public String getPrenom() {
+        return prenom;
+    }
+
+    public void setPrenom(String prenom) {
+        this.prenom = prenom;
+    }
+
+    public int getAge() {
+        return age;
+    }
+
+    public void setAge(int age) {
+        this.age = age;
+    }
+
+    public double getPoids() {
+        return poids;
+    }
+
+    public void setPoids(double poids) {
+        this.poids = poids;
+    }
+
+    public Sexe getSexe() {
+        return sexe;
+    }
+
+    public void setSexe(Sexe sexe) {
+        this.sexe = sexe;
+    }
+
+    public String getIdentifier() {
+        return identifier;
+    }
+
+    public void setIdentifier(String identifier) {
+        this.identifier = identifier;
+    }
+
+    public double getCurrentWeight() {
+        return currentWeight;
+    }
+
+    public void setCurrentWeight(double currentWeight) {
+        this.currentWeight = currentWeight;
+    }
+
+    public Double getHeight() {
+        return height;
+    }
+
+    public void setHeight(Double height) {
+        this.height = height;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getPhoneNumber() {
+        return phoneNumber;
+    }
+
+    public void setPhoneNumber(String phoneNumber) {
+        this.phoneNumber = phoneNumber;
+    }
+
+    public String getAddress() {
+        return address;
+    }
+
+    public void setAddress(String address) {
+        this.address = address;
+    }
+
+    public Timestamp getCreatedAt() {
+        return createdAt;
+    }
+
+    public void setCreatedAt(Timestamp createdAt) {
+        this.createdAt = createdAt;
+    }
+
+    public Timestamp getUpdatedAt() {
+        return updatedAt;
+    }
+
+    public void setUpdatedAt(Timestamp updatedAt) {
+        this.updatedAt = updatedAt;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+    public void setPassword(String password) {
+        this.password = password;
+    }
 }
