+++ src/main/java/Episante/back/Models/Question.java
@@ -11,22 +11,19 @@ public class Question {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String questionText;
-    private String questionType; // "number", "select", "text"
-    private String options; // Comma-separated options for select questions (e.g., "Male,Female,Other")
-    private String unit; // Unit for numeric input if any (e.g., "kg", "cm", "years")
 
-    public Question() {
-    }
+    public Question() {}
 
-    public Question(String questionText, String questionType, String options,String unit) {
+    public Question(String questionText) {
         this.questionText = questionText;
-        this.questionType = questionType;
-        this.options = options;
-        this.unit = unit;
     }
 
     public Long getId() {
         return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
     }
 
     public String getQuestionText() {
@@ -35,29 +32,5 @@ public class Question {
 
     public void setQuestionText(String questionText) {
         this.questionText = questionText;
-    }
-
-    public String getQuestionType() {
-        return questionType;
-    }
-
-    public void setQuestionType(String questionType) {
-        this.questionType = questionType;
-    }
-
-    public String getOptions() {
-        return options;
-    }
-
-    public void setOptions(String options) {
-        this.options = options;
-    }
-
-    public String getUnit() {
-        return unit;
-    }
-
-    public void setUnit(String unit) {
-        this.unit = unit;
     }
 }
